Created a weather data request manager which handles the logic of the need of refresh data list or not

+++ app/src/main/java/io/github/joaogouveia89/openweather/Constants.kt
@@ -1,5 +1,5 @@
 package io.github.joaogouveia89.openweather
 
 const val EARTH_RADIUS_KM = 6371
-const val KM_TO_RECALCULATE = 300
+const val KM_TO_RECALCULATE = 300.toDouble()
 const val DAYS_AGO_TO_UPDATE = 1
